stabilize CatalogAPI
