--- extensions/control-plane/api/management-api/catalog-api/src/main/java/org/eclipse/edc/connector/controlplane/api/management/catalog/v31alpha/CatalogApiV31AlphaController.java
@@ -1,52 +1,0 @@
-/*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.connector.controlplane.api.management.catalog.v31alpha;
-
-import jakarta.json.JsonObject;
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.container.AsyncResponse;
-import jakarta.ws.rs.container.Suspended;
-import org.eclipse.edc.connector.controlplane.api.management.catalog.BaseCatalogApiController;
-import org.eclipse.edc.connector.controlplane.services.spi.catalog.CatalogService;
-import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
-import org.eclipse.edc.validator.spi.JsonObjectValidatorRegistry;
-
-import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
-
-@Consumes(APPLICATION_JSON)
-@Produces(APPLICATION_JSON)
-@Path("/v3.1alpha/catalog")
-public class CatalogApiV31AlphaController extends BaseCatalogApiController implements CatalogApiV31alpha {
-    public CatalogApiV31AlphaController(CatalogService service, TypeTransformerRegistry transformerRegistry, JsonObjectValidatorRegistry validatorRegistry) {
-        super(service, transformerRegistry, validatorRegistry);
-    }
-
-    @POST
-    @Path("/request")
-    @Override
-    public void requestCatalogV31alpha(JsonObject request, @Suspended AsyncResponse response) {
-        requestCatalog(request, response);
-    }
-
-    @POST
-    @Path("dataset/request")
-    @Override
-    public void getDatasetV31alpha(JsonObject request, @Suspended AsyncResponse response) {
-        getDataset(request, response);
-    }
-}
