+++ extensions/control-plane/api/management-api/catalog-api/src/test/java/org/eclipse/edc/connector/controlplane/api/management/catalog/CatalogApiExtensionTest.java
@@ -15,7 +15,6 @@
 package org.eclipse.edc.connector.controlplane.api.management.catalog;
 
 import org.eclipse.edc.connector.controlplane.api.management.catalog.v3.CatalogApiV3Controller;
-import org.eclipse.edc.connector.controlplane.api.management.catalog.v31alpha.CatalogApiV31AlphaController;
 import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.validator.spi.JsonObjectValidatorRegistry;
@@ -57,6 +56,6 @@ class CatalogApiExtensionTest {
         extension.initialize(context);
 
         verify(webService).registerResource(any(), isA(CatalogApiV3Controller.class));
-        verify(webService).registerResource(any(), isA(CatalogApiV31AlphaController.class));
+        verify(webService).registerResource(any(), isA(CatalogApiV3Controller.class));
     }
 }
